requirements.txt
alembic
asyncpg
fastapi[all]
httpx
python-jose[cryptography]
passlib[bcrypt]
pytest
pytest-asyncio
python-dotenv
sqlalchemy
sqladmin

black
flake8
autoflake
isort
pyright